+++ extension/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v1.3.3 (In progress)
+## v1.3.4 (In progress)
 
 ### Extension